requirements.txt
streamlit>=1.51.0
anthropic>=0.73.0
browser-use-sdk>=0.1.0
python-dotenv>=1.0.0
daytona>=0.115.0
